pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mistral-ocr"
version = "0.1.0"
description = "A library for parsing PDF documents using Mistral's OCR API"
readme = "README.md"
authors = [
    {name = "Ravi Raina", email = "raviraina17@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "mistralai>=0.0.10",
    "python-dotenv>=1.0.0",
    "markdown>=3.5.1",
    "pillow>=10.0.0",
    "tqdm>=4.66.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/raviraina/mistral-ocr"
"Bug Tracker" = "https://github.com/raviraina/mistral-ocr/issues"

[project.scripts]
mistral-ocr = "mistral_ocr.parser:main"
mistral-ocr-batch = "mistral_ocr.utils:batch_process_pdfs"

[tool.setuptools]
packages = ["mistral_ocr"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 88
target-version = ["py37", "py38", "py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 